+++ src/shared/ui/input-phone/input-phone.stories.ts
